+++ rc.js
@@ -4,6 +4,10 @@
  */
 'use strict';
 
+const {join} = require('path');
+
+// Could make a simpler, more generic version for own repo
+
 module.exports = {
   overrides: [
     {
@@ -34,7 +38,17 @@ module.exports = {
       ],
       rules: {
         'no-console': 'off',
+        'import/no-anonymous-default-export': 'off',
         'node/no-process-env': 'off'
+      }
+    },
+    {
+      files: '*.md',
+      rules: {
+        'node/no-missing-require': ['error', {allowModules: [
+          // eslint-disable-next-line import/no-dynamic-require -- Meant to be flexible
+          require(join(process.cwd(), 'package.json')).name
+        ]}]
       }
     }
   ]