+++ sauron-node-script.js
@@ -1,5 +1,5 @@
 'use strict';
 
 module.exports = {
-  extends: ['./sauron-node.js', './_override-script-node.js']
+  extends: ['./sauron-node.js', './+script-node.js']
 };