+++ sauron-script.js
@@ -1,5 +1,5 @@
 'use strict';
 
 module.exports = {
-  extends: ['./sauron.js', './_override-script.js']
+  extends: ['./sauron.js', './+script.js']
 };